requirements.txt
gspread
kagglehub
pandas
requests
scikit-learn
apache-airflow
protobuf
cryptography
google-api-python-client
